+++ data_pipelines_cli/cli.py
@@ -1,3 +1,5 @@
+import sys
+
 import click
 
 from .cli_commands.clean import clean_command
@@ -8,18 +10,28 @@ from .cli_commands.init import init_command
 from .cli_commands.run import run_command
 from .cli_commands.template import list_templates_command
 from .cli_commands.test import test_command
+from .cli_utils import echo_error
+from .errors import DataPipelinesError
 
 
 @click.group()
-def cli() -> None:
+def _cli() -> None:
     pass
 
 
-cli.add_command(clean_command)
-cli.add_command(compile_project_command)
-cli.add_command(create_command)
-cli.add_command(deploy_command)
-cli.add_command(init_command)
-cli.add_command(run_command)
-cli.add_command(list_templates_command)
-cli.add_command(test_command)
+def cli() -> None:
+    try:
+        _cli()
+    except DataPipelinesError as err:
+        echo_error(f"CLI Error: {err.message}")
+        sys.exit(1)
+
+
+_cli.add_command(clean_command)
+_cli.add_command(compile_project_command)
+_cli.add_command(create_command)
+_cli.add_command(deploy_command)
+_cli.add_command(init_command)
+_cli.add_command(run_command)
+_cli.add_command(list_templates_command)
+_cli.add_command(test_command)